requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
replicate>=0.22.0
gunicorn>=21.2.0
django-cors-headers>=4.3.1
whitenoise>=6.6.0 